Update package name for releases going forward
