+++ app/src/main/AndroidManifest.xml
@@ -1,5 +1,5 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.devadvance.rootcloak" >
+    package="com.devadvance.rootcloak2" >
 
     <application
         android:allowBackup="true"
@@ -17,7 +17,7 @@
             android:value="Hides root from chosen apps." />
 
         <activity
-            android:name="com.devadvance.rootcloak.SettingsActivity"
+            android:name="com.devadvance.rootcloak2.SettingsActivity"
             android:label="@string/title_activity_settings" >
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
@@ -26,30 +26,30 @@
             </intent-filter>
         </activity>
         <activity
-            android:name="com.devadvance.rootcloak.CustomizeApps"
+            android:name="com.devadvance.rootcloak2.CustomizeApps"
             android:label="@string/title_activity_customize_apps"
-            android:parentActivityName="com.devadvance.rootcloak.SettingsActivity" >
+            android:parentActivityName="com.devadvance.rootcloak2.SettingsActivity" >
             <meta-data
                 android:name="android.support.PARENT_ACTIVITY"
-                android:value="com.devadvance.rootcloak.SettingsActivity" />
+                android:value="com.devadvance.rootcloak2.SettingsActivity" />
         </activity>
         
         <activity
-            android:name="com.devadvance.rootcloak.CustomizeKeywords"
+            android:name="com.devadvance.rootcloak2.CustomizeKeywords"
             android:label="@string/title_activity_customize_keywords"
-            android:parentActivityName="com.devadvance.rootcloak.SettingsActivity" >
+            android:parentActivityName="com.devadvance.rootcloak2.SettingsActivity" >
             <meta-data
                 android:name="android.support.PARENT_ACTIVITY"
-                android:value="com.devadvance.rootcloak.SettingsActivity" />
+                android:value="com.devadvance.rootcloak2.SettingsActivity" />
         </activity>
         
         <activity
-            android:name="com.devadvance.rootcloak.CustomizeCommands"
+            android:name="com.devadvance.rootcloak2.CustomizeCommands"
             android:label="@string/title_activity_customize_commands"
-            android:parentActivityName="com.devadvance.rootcloak.SettingsActivity" >
+            android:parentActivityName="com.devadvance.rootcloak2.SettingsActivity" >
             <meta-data
                 android:name="android.support.PARENT_ACTIVITY"
-                android:value="com.devadvance.rootcloak.SettingsActivity" />
+                android:value="com.devadvance.rootcloak2.SettingsActivity" />
         </activity>
     </application>
