+++ app/src/main/java/com/devadvance/rootcloak2/CustomizeApps.java
@@ -1,7 +1,9 @@
-package com.devadvance.rootcloak;
+package com.devadvance.rootcloak2;
 
 import java.util.Arrays;
+import java.util.HashMap;
 import java.util.HashSet;
+import java.util.List;
 import java.util.Set;
 
 import android.os.Bundle;
@@ -10,7 +12,8 @@ import android.app.AlertDialog;
 import android.content.DialogInterface;
 import android.content.SharedPreferences;
 import android.content.SharedPreferences.Editor;
-import android.content.res.Resources;
+import android.content.pm.ApplicationInfo;
+import android.content.pm.PackageManager;
 import android.view.Menu;
 import android.view.MenuItem;
 import android.view.View;
@@ -20,43 +23,38 @@ import android.widget.ListView;
 import android.support.v4.app.NavUtils;
 import android.preference.PreferenceActivity;
 
-public class CustomizeCommands extends PreferenceActivity {
+public class CustomizeApps extends PreferenceActivity {
 
     SharedPreferences sharedPref;
-    Set<String> commandSet;
-    String[] commandList;
-    boolean isFirstRunCommands;
+    Set<String> appSet;
+    String[] appList;
+    boolean isFirstRun;
 
     @SuppressLint("WorldReadableFiles")
     @SuppressWarnings("deprecation")
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_customize_commands);
+        setContentView(R.layout.activity_customize_apps);
         // Show the Up button in the action bar.
         setupActionBar();
 
 
         getPreferenceManager().setSharedPreferencesMode(MODE_WORLD_READABLE);
-        sharedPref = getSharedPreferences(Common.PREFS_COMMANDS, MODE_WORLD_READABLE);
+        sharedPref = getSharedPreferences(Common.PREFS_APPS, MODE_WORLD_READABLE);
 
         loadList();
-
-        Resources res = getResources();
-        new AlertDialog.Builder(this)
-        .setMessage(res.getString(R.string.command_instructions) + "\n\n" + res.getString(R.string.both_instructions2))
-        .setTitle(res.getString(R.string.important_title)).show();
 
     }
 
     public void onListItemClick( ListView parent, View v, int position, long id) {
         final int positionFinal = position;
-        new AlertDialog.Builder(CustomizeCommands.this)
-        .setTitle("Remove Command")
-        .setMessage("Are you sure you want to remove this command?")
+        new AlertDialog.Builder(CustomizeApps.this)
+        .setTitle("Remove App")
+        .setMessage("Are you sure you want to remove this app?")
         .setPositiveButton("Ok", new DialogInterface.OnClickListener() {
             public void onClick(DialogInterface dialog, int whichButton) {
-                removeCommand(positionFinal);
+                removeApp(positionFinal);
                 loadList();
             }
         }).setNegativeButton("Cancel", new DialogInterface.OnClickListener() {
@@ -79,7 +77,7 @@ public class CustomizeCommands extends PreferenceActivity {
     @Override
     public boolean onCreateOptionsMenu(Menu menu) {
         // Inflate the menu; this adds items to the action bar if it is present.
-        getMenuInflater().inflate(R.menu.customize_commands, menu);
+        getMenuInflater().inflate(R.menu.customize_apps, menu);
         return true;
     }
 
@@ -87,14 +85,45 @@ public class CustomizeCommands extends PreferenceActivity {
     public boolean onOptionsItemSelected(MenuItem item) {
         switch (item.getItemId()) {
         case android.R.id.home:
+            // This ID represents the Home or Up button. In the case of this
+            // activity, the Up button is shown. Use NavUtils to allow users
+            // to navigate up one level in the application structure. For
+            // more details, see the Navigation pattern on Android Design:
+            //
+            // http://developer.android.com/design/patterns/navigation.html#up-vs-back
+            //
             NavUtils.navigateUpFromSameTask(this);
             return true;
 
         case R.id.action_new:
+            final PackageManager pm = getPackageManager();
+            //get a list of installed apps.
+            final List<ApplicationInfo> packages = pm.getInstalledApplications(PackageManager.GET_META_DATA);
+            final String[] names = new String[packages.size()];
+            final HashMap<String, String> nameMap = new HashMap<String,String>();
+            int i = 0;
+            for (ApplicationInfo info : packages) {
+                //names[i] = info.packageName;
+                names[i] = (String)info.loadLabel(pm) + "\n(" + info.packageName + ")";
+                nameMap.put(names[i], info.packageName);
+                i++;
+            }
+            Arrays.sort(names);
+
+            new AlertDialog.Builder(this).setTitle("Add an App")
+            .setItems((CharSequence[]) names, new DialogInterface.OnClickListener() {
+                   public void onClick(DialogInterface dialog, int which) {
+                       savePref(nameMap.get(names[which]));
+                        loadList();
+               }
+        }).show();
+
+            return true;
+        case R.id.action_new_custom:
             final EditText input = new EditText(this);
-            new AlertDialog.Builder(CustomizeCommands.this)
-            .setTitle("Add Command")
-            .setMessage("Input the command:")
+            new AlertDialog.Builder(CustomizeApps.this)
+            .setTitle("Add App")
+            .setMessage("Input the app package name:")
             .setView(input)
             .setPositiveButton("Ok", new DialogInterface.OnClickListener() {
                 public void onClick(DialogInterface dialog, int whichButton) {
@@ -118,11 +147,11 @@ public class CustomizeCommands extends PreferenceActivity {
     }
 
     private void loadDefaults() {
-        commandSet = Common.DEFAULT_COMMAND_SET;
+        appSet = Common.DEFAULT_APPS_SET;
         Editor editor = sharedPref.edit();
-        editor.remove(Common.PACKAGE_NAME + Common.COMMAND_SET_KEY);
+        editor.remove(Common.PACKAGE_NAME + Common.APP_LIST_KEY);
         editor.commit();
-        editor.putStringSet(Common.PACKAGE_NAME + Common.COMMAND_SET_KEY, commandSet);
+        editor.putStringSet(Common.PACKAGE_NAME + Common.APP_LIST_KEY, appSet);
         editor.commit();
         editor.putBoolean(Common.PACKAGE_NAME + Common.FIRST_RUN_KEY, false);
         editor.commit();
@@ -130,8 +159,8 @@ public class CustomizeCommands extends PreferenceActivity {
     }
 
     private void loadDefaultsWithConfirm() {
-        AlertDialog.Builder builder = new AlertDialog.Builder(CustomizeCommands.this)
-        .setTitle("Reset commands to default?")
+        AlertDialog.Builder builder = new AlertDialog.Builder(CustomizeApps.this)
+        .setTitle("Reset apps to default?")
         .setPositiveButton(R.string.ok, new DialogInterface.OnClickListener() {
             public void onClick(DialogInterface dialog, int id) {
                 loadDefaults();
@@ -145,10 +174,10 @@ public class CustomizeCommands extends PreferenceActivity {
     }
 
     private void loadList() {
-        commandSet =  sharedPref.getStringSet(Common.PACKAGE_NAME + Common.COMMAND_SET_KEY, new HashSet<String>());
-        isFirstRunCommands = sharedPref.getBoolean(Common.PACKAGE_NAME + Common.FIRST_RUN_KEY, true);
-        if (isFirstRunCommands) {
-            if (commandSet.isEmpty()) {
+        appSet =  sharedPref.getStringSet(Common.PACKAGE_NAME + Common.APP_LIST_KEY, new HashSet<String>());
+        isFirstRun = sharedPref.getBoolean(Common.PACKAGE_NAME + Common.FIRST_RUN_KEY, true);
+        if (isFirstRun) {
+            if (appSet.isEmpty()) {
                 loadDefaults();
             }
             else {
@@ -157,22 +186,22 @@ public class CustomizeCommands extends PreferenceActivity {
                 editor.commit();
             }
         }
-        commandList = commandSet.toArray(new String[0]);
-        Arrays.sort(commandList);
+        appList = appSet.toArray(new String[0]);
+        Arrays.sort(appList);
 
         ArrayAdapter<String> adapter = new ArrayAdapter<String>(this,
-                android.R.layout.simple_list_item_1, commandList);
+                android.R.layout.simple_list_item_1, appList);
         // Bind to our new adapter.
         setListAdapter(adapter);
     }
 
     private void clearList() {
         final Editor editor = sharedPref.edit();
-        AlertDialog.Builder builder = new AlertDialog.Builder(CustomizeCommands.this)
-        .setTitle("Proceed to clear all commands?")
+        AlertDialog.Builder builder = new AlertDialog.Builder(CustomizeApps.this)
+        .setTitle("Proceed to clear all apps?")
         .setPositiveButton(R.string.ok, new DialogInterface.OnClickListener() {
             public void onClick(DialogInterface dialog, int id) {
-                editor.remove(Common.PACKAGE_NAME + Common.COMMAND_SET_KEY);
+                editor.remove(Common.PACKAGE_NAME + Common.APP_LIST_KEY);
                 editor.commit();
                 loadList();
             }
@@ -184,26 +213,26 @@ public class CustomizeCommands extends PreferenceActivity {
         }).show();
     }
 
-    private void savePref(String command) {
-        if (!(commandSet.contains(command))) {
-            commandSet.add(command);
+    private void savePref(String appName) {
+        if (!(appSet.contains(appName))) {
+            appSet.add(appName);
             Editor editor = sharedPref.edit();
-            editor.remove(Common.PACKAGE_NAME + Common.COMMAND_SET_KEY);
+            editor.remove(Common.PACKAGE_NAME + Common.APP_LIST_KEY);
             editor.commit();
-            editor.putStringSet(Common.PACKAGE_NAME + Common.COMMAND_SET_KEY, commandSet);
+            editor.putStringSet(Common.PACKAGE_NAME + Common.APP_LIST_KEY, appSet);
             editor.commit();
             editor.putBoolean(Common.PACKAGE_NAME + Common.FIRST_RUN_KEY, false);
             editor.commit();
         }
     }
 
-    private void removeCommand(int position) {
-        String tempName = commandList[position];
-        commandSet.remove(tempName);
+    private void removeApp(int position) {
+        String tempName = appList[position];
+        appSet.remove(tempName);
         Editor editor = sharedPref.edit();
-        editor.remove(Common.PACKAGE_NAME + Common.COMMAND_SET_KEY);
+        editor.remove(Common.PACKAGE_NAME + Common.APP_LIST_KEY);
         editor.commit();
-        editor.putStringSet(Common.PACKAGE_NAME + Common.COMMAND_SET_KEY, commandSet);
+        editor.putStringSet(Common.PACKAGE_NAME + Common.APP_LIST_KEY, appSet);
         editor.commit();
     }
