+++ app/src/main/java/com/devadvance/rootcloak2/CustomizeCommands.java
@@ -1,9 +1,7 @@
-package com.devadvance.rootcloak;
+package com.devadvance.rootcloak2;
 
 import java.util.Arrays;
-import java.util.HashMap;
 import java.util.HashSet;
-import java.util.List;
 import java.util.Set;
 
 import android.os.Bundle;
@@ -12,8 +10,7 @@ import android.app.AlertDialog;
 import android.content.DialogInterface;
 import android.content.SharedPreferences;
 import android.content.SharedPreferences.Editor;
-import android.content.pm.ApplicationInfo;
-import android.content.pm.PackageManager;
+import android.content.res.Resources;
 import android.view.Menu;
 import android.view.MenuItem;
 import android.view.View;
@@ -23,38 +20,43 @@ import android.widget.ListView;
 import android.support.v4.app.NavUtils;
 import android.preference.PreferenceActivity;
 
-public class CustomizeApps extends PreferenceActivity {
+public class CustomizeCommands extends PreferenceActivity {
 
     SharedPreferences sharedPref;
-    Set<String> appSet;
-    String[] appList;
-    boolean isFirstRun;
+    Set<String> commandSet;
+    String[] commandList;
+    boolean isFirstRunCommands;
 
     @SuppressLint("WorldReadableFiles")
     @SuppressWarnings("deprecation")
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_customize_apps);
+        setContentView(R.layout.activity_customize_commands);
         // Show the Up button in the action bar.
         setupActionBar();
 
 
         getPreferenceManager().setSharedPreferencesMode(MODE_WORLD_READABLE);
-        sharedPref = getSharedPreferences(Common.PREFS_APPS, MODE_WORLD_READABLE);
+        sharedPref = getSharedPreferences(Common.PREFS_COMMANDS, MODE_WORLD_READABLE);
 
         loadList();
+
+        Resources res = getResources();
+        new AlertDialog.Builder(this)
+        .setMessage(res.getString(R.string.command_instructions) + "\n\n" + res.getString(R.string.both_instructions2))
+        .setTitle(res.getString(R.string.important_title)).show();
 
     }
 
     public void onListItemClick( ListView parent, View v, int position, long id) {
         final int positionFinal = position;
-        new AlertDialog.Builder(CustomizeApps.this)
-        .setTitle("Remove App")
-        .setMessage("Are you sure you want to remove this app?")
+        new AlertDialog.Builder(CustomizeCommands.this)
+        .setTitle("Remove Command")
+        .setMessage("Are you sure you want to remove this command?")
         .setPositiveButton("Ok", new DialogInterface.OnClickListener() {
             public void onClick(DialogInterface dialog, int whichButton) {
-                removeApp(positionFinal);
+                removeCommand(positionFinal);
                 loadList();
             }
         }).setNegativeButton("Cancel", new DialogInterface.OnClickListener() {
@@ -77,7 +79,7 @@ public class CustomizeApps extends PreferenceActivity {
     @Override
     public boolean onCreateOptionsMenu(Menu menu) {
         // Inflate the menu; this adds items to the action bar if it is present.
-        getMenuInflater().inflate(R.menu.customize_apps, menu);
+        getMenuInflater().inflate(R.menu.customize_commands, menu);
         return true;
     }
 
@@ -85,45 +87,14 @@ public class CustomizeApps extends PreferenceActivity {
     public boolean onOptionsItemSelected(MenuItem item) {
         switch (item.getItemId()) {
         case android.R.id.home:
-            // This ID represents the Home or Up button. In the case of this
-            // activity, the Up button is shown. Use NavUtils to allow users
-            // to navigate up one level in the application structure. For
-            // more details, see the Navigation pattern on Android Design:
-            //
-            // http://developer.android.com/design/patterns/navigation.html#up-vs-back
-            //
             NavUtils.navigateUpFromSameTask(this);
             return true;
 
         case R.id.action_new:
-            final PackageManager pm = getPackageManager();
-            //get a list of installed apps.
-            final List<ApplicationInfo> packages = pm.getInstalledApplications(PackageManager.GET_META_DATA);
-            final String[] names = new String[packages.size()];
-            final HashMap<String, String> nameMap = new HashMap<String,String>();
-            int i = 0;
-            for (ApplicationInfo info : packages) {
-                //names[i] = info.packageName;
-                names[i] = (String)info.loadLabel(pm) + "\n(" + info.packageName + ")";
-                nameMap.put(names[i], info.packageName);
-                i++;
-            }
-            Arrays.sort(names);
-
-            new AlertDialog.Builder(this).setTitle("Add an App")
-            .setItems((CharSequence[]) names, new DialogInterface.OnClickListener() {
-                   public void onClick(DialogInterface dialog, int which) {
-                       savePref(nameMap.get(names[which]));
-                        loadList();
-               }
-        }).show();
-
-            return true;
-        case R.id.action_new_custom:
             final EditText input = new EditText(this);
-            new AlertDialog.Builder(CustomizeApps.this)
-            .setTitle("Add App")
-            .setMessage("Input the app package name:")
+            new AlertDialog.Builder(CustomizeCommands.this)
+            .setTitle("Add Command")
+            .setMessage("Input the command:")
             .setView(input)
             .setPositiveButton("Ok", new DialogInterface.OnClickListener() {
                 public void onClick(DialogInterface dialog, int whichButton) {
@@ -147,11 +118,11 @@ public class CustomizeApps extends PreferenceActivity {
     }
 
     private void loadDefaults() {
-        appSet = Common.DEFAULT_APPS_SET;
+        commandSet = Common.DEFAULT_COMMAND_SET;
         Editor editor = sharedPref.edit();
-        editor.remove(Common.PACKAGE_NAME + Common.APP_LIST_KEY);
+        editor.remove(Common.PACKAGE_NAME + Common.COMMAND_SET_KEY);
         editor.commit();
-        editor.putStringSet(Common.PACKAGE_NAME + Common.APP_LIST_KEY, appSet);
+        editor.putStringSet(Common.PACKAGE_NAME + Common.COMMAND_SET_KEY, commandSet);
         editor.commit();
         editor.putBoolean(Common.PACKAGE_NAME + Common.FIRST_RUN_KEY, false);
         editor.commit();
@@ -159,8 +130,8 @@ public class CustomizeApps extends PreferenceActivity {
     }
 
     private void loadDefaultsWithConfirm() {
-        AlertDialog.Builder builder = new AlertDialog.Builder(CustomizeApps.this)
-        .setTitle("Reset apps to default?")
+        AlertDialog.Builder builder = new AlertDialog.Builder(CustomizeCommands.this)
+        .setTitle("Reset commands to default?")
         .setPositiveButton(R.string.ok, new DialogInterface.OnClickListener() {
             public void onClick(DialogInterface dialog, int id) {
                 loadDefaults();
@@ -174,10 +145,10 @@ public class CustomizeApps extends PreferenceActivity {
     }
 
     private void loadList() {
-        appSet =  sharedPref.getStringSet(Common.PACKAGE_NAME + Common.APP_LIST_KEY, new HashSet<String>());
-        isFirstRun = sharedPref.getBoolean(Common.PACKAGE_NAME + Common.FIRST_RUN_KEY, true);
-        if (isFirstRun) {
-            if (appSet.isEmpty()) {
+        commandSet =  sharedPref.getStringSet(Common.PACKAGE_NAME + Common.COMMAND_SET_KEY, new HashSet<String>());
+        isFirstRunCommands = sharedPref.getBoolean(Common.PACKAGE_NAME + Common.FIRST_RUN_KEY, true);
+        if (isFirstRunCommands) {
+            if (commandSet.isEmpty()) {
                 loadDefaults();
             }
             else {
@@ -186,22 +157,22 @@ public class CustomizeApps extends PreferenceActivity {
                 editor.commit();
             }
         }
-        appList = appSet.toArray(new String[0]);
-        Arrays.sort(appList);
+        commandList = commandSet.toArray(new String[0]);
+        Arrays.sort(commandList);
 
         ArrayAdapter<String> adapter = new ArrayAdapter<String>(this,
-                android.R.layout.simple_list_item_1, appList);
+                android.R.layout.simple_list_item_1, commandList);
         // Bind to our new adapter.
         setListAdapter(adapter);
     }
 
     private void clearList() {
         final Editor editor = sharedPref.edit();
-        AlertDialog.Builder builder = new AlertDialog.Builder(CustomizeApps.this)
-        .setTitle("Proceed to clear all apps?")
+        AlertDialog.Builder builder = new AlertDialog.Builder(CustomizeCommands.this)
+        .setTitle("Proceed to clear all commands?")
         .setPositiveButton(R.string.ok, new DialogInterface.OnClickListener() {
             public void onClick(DialogInterface dialog, int id) {
-                editor.remove(Common.PACKAGE_NAME + Common.APP_LIST_KEY);
+                editor.remove(Common.PACKAGE_NAME + Common.COMMAND_SET_KEY);
                 editor.commit();
                 loadList();
             }
@@ -213,26 +184,26 @@ public class CustomizeApps extends PreferenceActivity {
         }).show();
     }
 
-    private void savePref(String appName) {
-        if (!(appSet.contains(appName))) {
-            appSet.add(appName);
+    private void savePref(String command) {
+        if (!(commandSet.contains(command))) {
+            commandSet.add(command);
             Editor editor = sharedPref.edit();
-            editor.remove(Common.PACKAGE_NAME + Common.APP_LIST_KEY);
+            editor.remove(Common.PACKAGE_NAME + Common.COMMAND_SET_KEY);
             editor.commit();
-            editor.putStringSet(Common.PACKAGE_NAME + Common.APP_LIST_KEY, appSet);
+            editor.putStringSet(Common.PACKAGE_NAME + Common.COMMAND_SET_KEY, commandSet);
             editor.commit();
             editor.putBoolean(Common.PACKAGE_NAME + Common.FIRST_RUN_KEY, false);
             editor.commit();
         }
     }
 
-    private void removeApp(int position) {
-        String tempName = appList[position];
-        appSet.remove(tempName);
+    private void removeCommand(int position) {
+        String tempName = commandList[position];
+        commandSet.remove(tempName);
         Editor editor = sharedPref.edit();
-        editor.remove(Common.PACKAGE_NAME + Common.APP_LIST_KEY);
+        editor.remove(Common.PACKAGE_NAME + Common.COMMAND_SET_KEY);
         editor.commit();
-        editor.putStringSet(Common.PACKAGE_NAME + Common.APP_LIST_KEY, appSet);
+        editor.putStringSet(Common.PACKAGE_NAME + Common.COMMAND_SET_KEY, commandSet);
         editor.commit();
     }
