+++ app/src/main/java/com/devadvance/rootcloak2/CustomizeKeywords.java
@@ -1,4 +1,4 @@
-package com.devadvance.rootcloak;
+package com.devadvance.rootcloak2;
 
 import java.util.Arrays;
 import java.util.HashSet;
