+++ app/src/main/java/com/devadvance/rootcloak2/RootCloak.java
@@ -1,4 +1,4 @@
-package com.devadvance.rootcloak;
+package com.devadvance.rootcloak2;
 
 import static de.robv.android.xposed.XposedHelpers.*;
